Refactor RyuJit
Set up the consumption of RyuJit classes so that we can import CorInfoImpl.cs from multiple projects with different `#ifdef` settings for conditional compilation.

* In CoreCLR ready-to-run compilation mode, we want the Jit interface implementation to know about ReadyToRunCodegenCompilation so that it can ask for nodes / services that are ready-to-run specific. We do not want to leak that stuff into the main ILCompiler.Compiler project, since vanilla CoreRT shouldn't have to know about such details.
* Revert the class visibility of several JitInterface classes. We made some public to speed up the prototyping previousl.
* Make the type visibility uniform in `CorInfoTypes.cs`. There was a mixture of public / internal structs / enums. I chose to make them all public so this file can be included in ILCompiler.RyuJit and its contents referenced from consuming libraries.
* Define `READY_TO_RUN` when building `ILCompiler.ReadyToRun` so we can keep the `ReadyToRunCodegenCompilation` under `#ifdef`
* Move Jit-specific classes to `ILCompiler.RyuJit`: `RyuJitNodeFactory`, `RyuJitCompilation`, `RyuJitCompilationBuilder`, `MethodCodeNode`, `IMethodCodeNode`
