+++ src/ILCompiler/ILCompiler.sln
@@ -31,6 +31,7 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ILCompiler.Build.Tasks", "..\ILCompiler.Build.Tasks\src\ILCompiler.Build.Tasks.csproj", "{44DE7F7B-AD73-40EA-87CC-554F2F57C38A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ILCompiler.ReadyToRun", "..\ILCompiler.ReadyToRun\src\ILCompiler.ReadyToRun.csproj", "{5FB7EFC5-2E58-4A65-A611-C9B68CC6A78D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ILCompiler.RyuJit", "..\ILCompiler.RyuJit\src\ILCompiler.RyuJit.csproj", "{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
@@ -405,6 +406,34 @@ Global
 		{5FB7EFC5-2E58-4A65-A611-C9B68CC6A78D}.Release|x64.Build.0 = Release|x64
 		{5FB7EFC5-2E58-4A65-A611-C9B68CC6A78D}.Release|x86.ActiveCfg = Release|x86
 		{5FB7EFC5-2E58-4A65-A611-C9B68CC6A78D}.Release|x86.Build.0 = Release|x86
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|arm.ActiveCfg = Debug|arm
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|arm.Build.0 = Debug|arm
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|arm64.ActiveCfg = Debug|arm64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|arm64.Build.0 = Debug|arm64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|armel.ActiveCfg = Debug|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|armel.Build.0 = Debug|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|wasm.ActiveCfg = Debug|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|wasm.Build.0 = Debug|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|x64.ActiveCfg = Debug|x64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|x64.Build.0 = Debug|x64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|x86.ActiveCfg = Debug|x86
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Debug|x86.Build.0 = Debug|x86
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|arm.ActiveCfg = Release|arm
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|arm.Build.0 = Release|arm
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|arm64.ActiveCfg = Release|arm64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|arm64.Build.0 = Release|arm64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|armel.ActiveCfg = Release|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|armel.Build.0 = Release|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|wasm.ActiveCfg = Release|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|wasm.Build.0 = Release|Any CPU
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|x64.ActiveCfg = Release|x64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|x64.Build.0 = Release|x64
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|x86.ActiveCfg = Release|x86
+		{69CB6CF1-9B6C-4617-A8B7-F0089CE474FB}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
