+++ src/ILCompiler/desktop/desktop.csproj
@@ -88,6 +88,10 @@
       <Project>{971ae7e7-08c6-48b4-902a-63851e6dac66}</Project>
       <Name>ILCompiler.CppCodegen</Name>
     </ProjectReference>
+    <ProjectReference Include="..\..\ILCompiler.RyuJit\src\ILCompiler.RyuJit.csproj">
+      <Project>{69cb6cf1-9b6c-4617-a8b7-f0089ce474fb}</Project>
+      <Name>ILCompiler.RyuJit</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\ILCompiler.TypeSystem\src\ILCompiler.TypeSystem.csproj">
       <Project>{1a9df196-43a9-44bb-b2c6-d62aa56b0e49}</Project>
       <Name>ILCompiler.TypeSystem</Name>
